+++ include/Camera.h
@@ -63,7 +63,7 @@ public:
     // constructor with vectors
     Camera(glm::vec3 position = glm::vec3(0.0f, 0.0f, -20.0f), glm::vec3 up = glm::vec3(0.0f, 1.0f, 0.0f), float yaw = YAW, float pitch = PITCH) : Front(glm::vec3(0.0f, 0.0f, -1.0f)), MovementSpeed(SPEED), MouseSensitivity(SENSITIVITY), Zoom(ZOOM)
     {
-        spdlog::warn("camera constructor");
+        //spdlog::warn("camera constructor");
 
         Position = position;
         WorldUp = up;
@@ -76,7 +76,7 @@ public:
     // constructor with scalar values
     Camera(float posX, float posY, float posZ, float upX, float upY, float upZ, float yaw, float pitch) : Front(glm::vec3(0.0f, 0.0f, -1.0f)), MovementSpeed(SPEED), MouseSensitivity(SENSITIVITY), Zoom(ZOOM)
     {
-        spdlog::warn("camera constructor");
+        //spdlog::warn("camera constructor");
 
         Position = glm::vec3(posX, posY, posZ);
         WorldUp = glm::vec3(upX, upY, upZ);
@@ -87,7 +87,7 @@ public:
     }
 
     ~Camera() {
-        spdlog::warn("camera destructor");
+        spdlog::error("camera destructor");
     }
 
     void rotate(float xdeg, float ydeg, GLboolean constrainPitch = true)