+++ src/Engine.cpp
@@ -215,7 +215,7 @@ namespace Engine {
         for(Object3D* object: allObjects) {
 //                object->Draw();
 //            object->_transform.computeModelMatrix();
-            Entity test(object->_model);
+            Entity test(*object->_model);
             test.transform.setLocalPosition(object->_transform._position);
             test.transform.setLocalScale(object->_transform._scale);
             test.transform.setLocalRotation(object->_transform._rotation);